Extras moved to /extra folder and their own modules
