+++ examples/full-app-gourmet/go.mod
@@ -1,11 +1,12 @@
 module github.com/go-fuego/fuego/examples/full-app-gourmet
 
-go 1.22.2
+go 1.22.5
 
 require (
 	github.com/a-h/templ v0.2.747
 	github.com/go-chi/chi/v5 v5.1.0
 	github.com/go-fuego/fuego v0.14.0-rc2
+	github.com/go-fuego/fuego/extra/markdown v0.0.0-00010101000000-000000000000
 	github.com/golang-jwt/jwt/v5 v5.2.1
 	github.com/golang-migrate/migrate/v4 v4.17.1
 	github.com/google/uuid v1.6.0
@@ -57,3 +58,7 @@ require (
 )
 
 replace github.com/go-fuego/fuego => ../..
+
+replace github.com/go-fuego/fuego/extra/markdown => ../../extra/markdown
+
+replace github.com/go-fuego/fuego/middleware/cache => ../../middleware/cache
