+++ extra/markdown/go.mod
@@ -1,0 +1,5 @@
+module github.com/go-fuego/fuego/extra/markdown
+
+go 1.22.5
+
+require github.com/gomarkdown/markdown v0.0.0-20240626202925-2eda941fd024
