+++ go.work
@@ -1,4 +1,4 @@
-go 1.22.4
+go 1.22.5
 
 use (
 	.
@@ -9,5 +9,6 @@ use (
 	./examples/full-app-gourmet
 	./examples/hello-world
 	./examples/openapi
+	./extra/markdown
 	./testing-from-outside
 )
